% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diveparams.R
\name{calculate_diveparams}
\alias{calculate_diveparams}
\title{Calculates diving parameters for the complete trip}
\usage{
calculate_diveparams(TDR_dives = TDR_dives)
}
\arguments{
\item{TDR_dives}{is a data frame with all the dives that ocurred during the trip}
}
\value{
dataframe with dive parameters
}
\description{
Calculates diving parameters for the complete trip
}
\examples{
dive_parameters<-calculate_diveparams(TDR_dives)
}
