% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_zero.R
\name{correct_zero}
\alias{correct_zero}
\title{Corrects zero.}
\usage{
correct_zero(
  TDR_trip = TDR_trip,
  depth_column = depth_column,
  extra_correction = extra_correction
)
}
\arguments{
\item{TDR_trip}{is a data frame with a column with depth data. Escribe el nombre de tu data frame aqui, debe contener una columna con datos de profundidad.}

\item{depth_column}{is the name of the column that contains the depth data. Escribe aqui entre comillas el nombre de la columna con datos de profundidad.}

\item{extra_correction}{write 0 if no extra correction is needer, otherwise add a number here. Escribe 0 si no es necesario corregir, de otra manera pon un numero aqui.}
}
\value{
returns a data frame with a column corrected_depth, with the corrected data.
}
\description{
Corrects zero.
}
\examples{
TDR_corrected<-correct_zero(TDR_trip = TDR_trip, depth_column='Pressure',extra_correction=-0.65)
}
