% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_trip.R
\name{cut_trip}
\alias{cut_trip}
\title{Function to cut periods of the TDR that correspond to time at sea}
\usage{
cut_trip(
  data = data,
  timeformat = timeformat,
  trip_start = trip_start,
  trip_end = trip_end
)
}
\arguments{
\item{data}{your TDR data with at least these two columns: tStamp and Pressure}

\item{timeformat}{the format of your data in your TDR data tStamp}

\item{trip_start}{The period of time when the trip started. This data can be deduced from field observations or obtained from GPS.}

\item{trip_end}{The period of time when the trip finished. This data can be deduced from field observations or obtained from GPS.}
}
\value{
Returns a object with less observations than in the original data. Only the periods between the data provided are returned.
}
\description{
Function to cut periods of the TDR that correspond to time at sea
}
\examples{
timeformat<-"\%d-\%m-\%Y \%H:\%M:\%S"
trip_start<-"30-11-2018 20:43:24"
trip_end<-"01-12-2018 20:16:19"
TDR_trip<-cut_trip(data=TDR_pressure,timeformat=timeformat,trip_start=trip_start,trip_end=trip_end)
}
