% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pressure.R
\name{extract_pressure}
\alias{extract_pressure}
\title{EN: Extracts pressure data from raw time and depth data. ES: Extrae data de presion desde dispositivos TDR}
\usage{
extract_pressure(data = data, row_start = row_start, row_end = row_end)
}
\arguments{
\item{data}{EN: is the raw data as imported from your device. ES:son los datos crudos descargados de tu dispositivo}

\item{row_start}{EN: is the name of the row where the pressure data starts to be recorded. ES: es la nombre de la fila donde empieza a tomar los datos de presion}

\item{row_end}{EN: is the name of the row where the pressure data finishes to be recorded. ES: es la nombre de la fila donde termina de tomar los datos de presion}
}
\value{
EN: it will return a object named TDR_pressure. ES: agrega un objeto llamado TDR_pressure al environment
}
\description{
EN: Extracts pressure data from raw time and depth data. ES: Extrae data de presion desde dispositivos TDR
}
\examples{
TDR_raw<-TDR_raw
TDR_pressure<-extract_pressure(data=TDR_raw, row_start='Data Block 1', row_end = 'Data Block 2')
}
