% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trip.R
\name{extract_trip}
\alias{extract_trip}
\title{Function to extract periods of the TDR that correspond to time at sea}
\usage{
extract_trip(
  data = data,
  timeformat = timeformat,
  trip_start = trip_start,
  trip_end = trip_end
)
}
\arguments{
\item{data}{your TDR data with at least these two columns: tStamp and Pressure}

\item{timeformat}{the format of your data in your TDR data tStamp}

\item{trip_start}{The period of time when the trip started. This data can be deduced from field observations or obtained from GPS.}

\item{trip_end}{The period of time when the trip finished. This data can be deduced from field observations or obtained from GPS.}
}
\value{
Returns a object with less observations than in the original data. Only the periods between the data provided are returned.
}
\description{
Function to extract periods of the TDR that correspond to time at sea
}
\examples{
t_format<-"\%d-\%m-\%Y \%H:\%M:\%S"
t_start<-"30-11-2018 20:43:24"
t_end<-"01-12-2018 20:16:19"
TDR_trip<-extract_trip(data=TDR_pressure,timeformat=t_format,trip_start=t_start,trip_end=t_end)
}
