% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_dives.R
\name{identify_dives}
\alias{identify_dives}
\title{This functions identify dives according to a parameter. Esta funcion identifica buceos a partir de cierta profundidad/}
\usage{
identify_dives(
  TDR_corrected = TDR_corrected,
  depth_column = depth_column,
  real_dives = real_dives
)
}
\arguments{
\item{TDR_corrected}{is the dataframe with depth already zero-corrected. Es el data frame con la profundidad corregida.}

\item{depth_column}{is a numeric column where the depth information is. Es el nombre de la columna donde esta la informacion de profundidad.}

\item{real_dives}{a number when dives are considered as real dives. Un numero de los cuando los buceos son considerados buceos reales.}
}
\value{
returns a data frame with each dive duration, maximum depth, mean and standard deviation depth
}
\description{
This functions identify dives according to a parameter. Esta funcion identifica buceos a partir de cierta profundidad/
}
\examples{
table_dives<-identify_dives(TDR_corrected=TDR_corrected,real_dives=3,depth_column='corrected_depth')
}
