% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_depth.R
\name{plot_depth}
\alias{plot_depth}
\title{plot_depth: Creates a plot with the diving profile. Adds a red line for the zero. Crea un grafico con el perfil de buceos. Marca el cero con una linea roja.}
\usage{
plot_depth(
  TDR_trip = TDR_trip,
  depth_column = depth_column,
  time_column = time_column
)
}
\arguments{
\item{TDR_trip}{is a data frame with a colum including depth and a column including time (in POSIXct format)}

\item{depth_column}{must be a numeric value}

\item{time_column}{must be a date and time (in POSIXct format) (\%Y-\%m-d \%H:\%M:\%S)}
}
\value{
returns a ggplot which stetics can be change using the ggplot2 sintaxis
}
\description{
plot_depth: Creates a plot with the diving profile. Adds a red line for the zero. Crea un grafico con el perfil de buceos. Marca el cero con una linea roja.
}
\examples{
plot_depth(TDR_trip = TDR_trip,depth_column='Pressure',time_column='daytime')
}
