% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDR_corrected.R
\docType{data}
\name{TDR_corrected}
\alias{TDR_corrected}
\title{Corrected data from TDRs}
\format{
A rda data with one column and 42388 rows
}
\usage{
data(TDR_corrected)
}
\description{
EN:
The data includes pressure data obtained from a Humbolt Penguin at Tilgo, Chile
The information was collected during the breeding season 2018.
It contains 42388 obs from 6 variables: tStamp, Pressure, numeric_sequence, daytime, depth and corrected_depth
ES:
Los datos incluyen datos crudos de un pinguino de Humbold muestreado en Isla Tilgo, Chile
La information fue colectada durante la temporada reproductiva 2018.
Contiene 42388 observaciones de 6 variables: tStamp, Pressure, numeric_sequence, daytime, depth and corrected_depth
}
\examples{
data(TDR_corrected)
}
\references{
Lerma et al. 2021
}
\keyword{datasets}
