% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDR_pressure.R
\docType{data}
\name{TDR_pressure}
\alias{TDR_pressure}
\title{Pressure data from TDRS}
\format{
A rda data with one column and 255600 rows
}
\usage{
data(TDR_pressure)
}
\description{
EN:
The data includes pressure data obtained from a Humbolt Penguin at Tilgo, Chile
The information was collected during the breeding season 2018.
It contains 255600 obs from 3 variables: tStamp, Pressure, numeric_sequence.
ES:
Los datos incluyen datos crudos de un pinguino de Humbold muestreado en Isla Tilgo, Chile
La information fue colectada durante la temporada reproductiva 2018.
Contiene 255600 observaciones de 3 variables: tStamp, Pressure, numeric_sequence.
}
\examples{
data(TDR_pressure)
}
\references{
Lerma et al. 2021
}
\keyword{datasets}
