% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDR_raw.R
\docType{data}
\name{TDR_raw}
\alias{TDR_raw}
\title{Raw data from a TDR (time and depth recorder) of a Humboldt penguin}
\format{
A rda data with one column and 264197 rows
}
\usage{
data(TDR_raw)
}
\description{
The data includes raw data obtained from a Humbolt Penguin at Tilgo, Chile
The information was collected during the breeding season 2018.
It contains 264197 obs from 1 variable.
}
\examples{
data(TDR_raw)
}
\references{
Lerma et al. 2021
}
\keyword{datasets}
