% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDR_trip.R
\docType{data}
\name{TDR_trip}
\alias{TDR_trip}
\title{TDR_trip contains information only from when the bird was at sea}
\format{
A rda data with one column and 255600 rows
}
\usage{
data(TDR_trip)
}
\description{
EN:
The data includes pressure data obtained from a Humbolt Penguin at Tilgo, Chile
The information was collected during the breeding season 2018.
It contains 42388 obs from 4 variables: tStamp, Pressure, numeric_sequence, daytime.
ES:
Los datos incluyen datos crudos de un pinguino de Humbold muestreado en Isla Tilgo, Chile
La information fue colectada durante la temporada reproductiva 2018.
Contiene 42388 observaciones de 4 variables: tStamp, Pressure, numeric_sequence y daytime.
}
\examples{
data(TDR_trip)
}
\references{
Lerma et al. 2021
}
\keyword{datasets}
