% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDR_raw.R
\docType{data}
\name{TDR_raw}
\alias{TDR_raw}
\title{EN: Raw data from a time and depth recorder of a Humboldt penguin. ES: Datos crudos de TDR en Pinguinos de Humboldt}
\format{
A rda data with one column and 264197 rows
}
\usage{
data(TDR_raw)
}
\description{
EN:
The data includes raw data obtained from a Humbolt Penguin at Tilgo, Chile
The information was collected during the breeding season 2018.
It contains 264197 obs from 1 variable.
Data is provided as it is imported from the device.
ES:
Los datos incluyen datos crudos de un pinguino de Humbold muestreado en Isla Tilgo, Chile
La information fue colectada durante la temporada reproductiva 2018.
Contiene 264197 observaciones de 1 variable.
Es un ejemplo de como los datos son descargados del dispositivo.
}
\examples{
data(TDR_raw)
}
\references{
Lerma et al. 2021
}
\keyword{datasets}
